Some fixes

+++ CurlingApp/CurlingApp/CoreData/CURGameInfo.h
@@ -15,8 +15,8 @@
 @property (nullable, nonatomic, copy) NSString *teamNameSecond;
 @property (nullable, nonatomic, copy) NSString *hashLink;
 @property (nonatomic) BOOL isFirstTeamColorRed;
-@property (nonatomic) int32_t numberOfEnds;
-@property (nonatomic) int32_t firstTeamScore;
-@property (nonatomic) int32_t secondTeamScore;
+@property (nonatomic) NSInteger numberOfEnds;
+@property (nonatomic) NSInteger firstTeamScore;
+@property (nonatomic) NSInteger secondTeamScore;
 
 @end
